+++ CHANGELOG.md
@@ -39,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- [#618](https://github.com/alleslabs/celatone-frontend/pull/618) Add save account function in account detail page
+- [#620](https://github.com/alleslabs/celatone-frontend/pull/620) Add hex address in save account feature in initia
+- [#618](https://github.com/alleslabs/celatone-frontend/pull/618) Add save account feature in account detail page
 - [#616](https://github.com/alleslabs/celatone-frontend/pull/616) Add table for saved accounts and add save, remove and edit modal
 - [#613](https://github.com/alleslabs/celatone-frontend/pull/613) Add saved accounts modal ui
 - [#611](https://github.com/alleslabs/celatone-frontend/pull/611) Add saved accounts page